LayoutTests/fast/css/css-typed-om/style-property-map-set-CSSMathSum-value-expected.txt

Checks that calling StylePropertyMap.set() with a CSSMathSum value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.attributeStyleMap.set('width', new CSSMathSum(new CSSUnitValue(1, 'percent'), new CSSUnitValue(2, 'percent'))) did not throw exception.
PASS target.style.width == 'calc(3%)' || target.style.width == 'calc(1% + 2%)' is true
PASS successfullyParsed is true

TEST COMPLETE
